requirements.txt
google-cloud-logging
Flask
xialib-firestore
xialib-gcs
xialib
gunicorn
pyinsight
google-cloud-firestore